Use new draft invoice generator as default
And remove diffing tool
