+++ service/src/main/kotlin/fi/espoo/evaka/EspooConfig.kt
@@ -9,11 +9,9 @@ import fi.espoo.evaka.emailclient.EvakaEmailMessageProvider
 import fi.espoo.evaka.emailclient.IEmailMessageProvider
 import fi.espoo.evaka.invoicing.integration.EspooInvoiceIntegrationClient
 import fi.espoo.evaka.invoicing.integration.InvoiceIntegrationClient
-import fi.espoo.evaka.invoicing.service.DefaultDraftInvoiceGenerator
 import fi.espoo.evaka.invoicing.service.EspooIncomeTypesProvider
 import fi.espoo.evaka.invoicing.service.EspooInvoiceProducts
 import fi.espoo.evaka.invoicing.service.IncomeTypesProvider
-import fi.espoo.evaka.invoicing.service.InvoiceGenerator
 import fi.espoo.evaka.invoicing.service.InvoiceProductProvider
 import fi.espoo.evaka.shared.FeatureConfig
 import fi.espoo.evaka.shared.db.DevDataInitializer
@@ -73,10 +71,6 @@ class EspooConfig {
 
     @Bean
     fun invoiceProductsProvider(): InvoiceProductProvider = EspooInvoiceProducts.Provider()
-
-    @Bean
-    fun invoiceGenerator(productProvider: InvoiceProductProvider): InvoiceGenerator =
-        InvoiceGenerator(DefaultDraftInvoiceGenerator(productProvider))
 
     @Bean
     fun featureConfig(): FeatureConfig = FeatureConfig(
